Fix bug with deprecation notice (#2367)
'Implicit conversion from float 1.2 to int loses precision in .. line 240'.
This make the type conversion explicit.

+++ src/Credentials/InstanceProfileProvider.php
@@ -237,7 +237,7 @@ class InstanceProfileProvider
             $isRetryable = false;
         }
         if ($isRetryable && $this->attempts < $this->retries) {
-            sleep(pow(1.2, $this->attempts));
+            sleep((int) pow(1.2, $this->attempts));
         } else {
             throw new CredentialsException($message);
         }
